+++ alma/controllers/hook/displayPayment.php
@@ -31,6 +31,7 @@ include_once _PS_MODULE_DIR_ . 'alma/includes/AlmaClient.php';
 include_once _PS_MODULE_DIR_ . 'alma/includes/AlmaSettings.php';
 include_once _PS_MODULE_DIR_ . 'alma/includes/AlmaEligibilityHelper.php';
 include_once _PS_MODULE_DIR_ . 'alma/includes/functions.php';
+include_once _PS_MODULE_DIR_ . 'alma/includes/CartData.php';
 
 class AlmaDisplayPaymentController extends AlmaProtectedHookController
 {
@@ -42,6 +43,12 @@ class AlmaDisplayPaymentController extends AlmaProtectedHookController
         $error = false;
         if (!$eligibility) {
             $error = true;
+        }
+
+        // Check if some products in cart are in the excludes listing
+        $diff = CartData::getCartExclusion($params['cart']); 
+        if(!empty($diff)){
+            return false;
         }
 
         if (isset($eligibility) && $eligibility->isEligible) {