+++ alma/controllers/hook/getContent.php
@@ -53,6 +53,7 @@ class AlmaGetContentController extends AlmaAdminHookController
             $showEligibility = Tools::getValue('ALMA_SHOW_ELIGIBILITY_MESSAGE_ON') == '1';
             $eligibleMsg = Tools::getValue('ALMA_IS_ELIGIBLE_MESSAGE');
             $nonEligibleMsg = Tools::getValue('ALMA_NOT_ELIGIBLE_MESSAGE');
+            $nonEligibleCategoriesMsg = Tools::getValue('ALMA_NOT_ELIGIBLE_CATEGORIES');
 
             if (empty($title) || empty($description) ||
                 ($showEligibility && (empty($eligibleMsg) || empty($nonEligibleMsg)))) {
@@ -70,6 +71,7 @@ class AlmaGetContentController extends AlmaAdminHookController
             AlmaSettings::updateValue('ALMA_SHOW_ELIGIBILITY_MESSAGE', $showEligibility);
             AlmaSettings::updateValue('ALMA_IS_ELIGIBLE_MESSAGE', $eligibleMsg);
             AlmaSettings::updateValue('ALMA_NOT_ELIGIBLE_MESSAGE', $nonEligibleMsg);
+            AlmaSettings::updateValue('ALMA_NOT_ELIGIBLE_CATEGORIES', $nonEligibleCategoriesMsg);
 
             $idStateRefund = Tools::getValue('ALMA_STATE_REFUND');
             AlmaSettings::updateValue('ALMA_STATE_REFUND', $idStateRefund);
@@ -591,6 +593,49 @@ class AlmaGetContentController extends AlmaAdminHookController
             ),
         );
 
+
+
+        // Exclusion
+        $tpl = $this->context->smarty->createTemplate(
+            "{$this->module->local_path}views/templates/hook/excludedCategories.tpl"
+        );
+
+        $excludedCategoryNames = AlmaSettings::getExcludedCategoryNames();
+
+        $tpl->assign(array(
+            'excludedCategories' => count($excludedCategoryNames) > 0
+				? implode(', ', $excludedCategoryNames)
+				: $this->module->l('No excluded categories', 'getContent'),
+            'excludedLink' => $this->context->link->getAdminLink('AdminAlmaCategories'),
+        ));
+        $excludedForm = array(
+            'form' => array(
+                'legend' => array(
+                    'title' => $this->module->l('Excluded categories', 'getContent'),
+                    'image' => $iconPath,
+                ),
+                'input' => array(
+                    array(
+                        'name' => null,
+                        'label' => null,
+                        'type' => 'html',
+                        // PrestaShop won't detect the string if the call to `l` is multiline
+                        'desc' => $tpl->fetch(),
+                    ),
+                    array(
+                        'name' => 'ALMA_NOT_ELIGIBLE_CATEGORIES',
+                        'label' => $this->module->l('Excluded categories non-eligibility message ', 'getContent'),
+                        // PrestaShop won't detect the string if the call to `l` is multiline
+                        'desc' => $this->module->l('Message displayed on an excluded product page or on the cart page if it contains an excluded product.', 'getContent'),
+                        'type' => 'text',
+                        'size' => 75,
+                        'required' => false,
+                    ),
+                ),
+                'submit' => array('title' => $this->module->l('Save'), 'class' => 'btn btn-default pull-right'),
+            ),
+        );
+
         $refundStateForm = array(
             'form' => array(
                 'legend' => array(
@@ -681,7 +726,7 @@ class AlmaGetContentController extends AlmaAdminHookController
                 $fieldsForms[] = $pnxConfigForm;
             }
 
-            $fieldsForms = array_merge($fieldsForms, array($orderConfirmationForm, $apiConfigForm, $debugForm));
+            $fieldsForms = array_merge($fieldsForms, array($orderConfirmationForm, $excludedForm, $apiConfigForm, $debugForm));
         }
 
         $helper = new HelperForm();
@@ -716,6 +761,7 @@ class AlmaGetContentController extends AlmaAdminHookController
             'ALMA_ACTIVATE_LOGGING_ON' => (bool)AlmaSettings::canLog(),
             'ALMA_STATE_REFUND' => AlmaSettings::getRefundState(),
             'ALMA_STATE_REFUND_ENABLED_ON' => AlmaSettings::isRefundEnabledByState(),
+            'ALMA_NOT_ELIGIBLE_CATEGORIES' => AlmaSettings::getNonEligibilityCategoriesMessage(),
             '_api_only' => true,
         );